Commenting, refactoring
+ Added English comments to the code
+ Light refactoring of some classes
