+++ ClassFinder/pom.xml
@@ -22,6 +22,11 @@
           </archive>
         </configuration>
       </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-javadoc-plugin</artifactId>
+        <version>2.9.1</version>
+      </plugin>
     </plugins>
   </build>
 </project>
