+++ ClassFinder/src/main/java/jars/search/DeleteListener.java
@@ -3,9 +3,20 @@ package jars.search;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+/**
+ * This class implements an action listener for the Delete button
+ */
 final class DeleteListener implements ActionListener {
+	
+	//--------------------------------------------------
+	// Class properties 
+	
+	// Reference to the search panel
 	private SearchPanel panel;
 
+	//--------------------------------------------------
+	// Class methods 
+	
 	public DeleteListener(SearchPanel searchPanel) {
 		panel = searchPanel;
 	}
