+++ ClassFinder/src/main/java/jars/search/FileListModel.java
@@ -5,13 +5,28 @@ import java.util.Collections;
 import java.util.List;
 import javax.swing.AbstractListModel;
 
+/**
+ * This class implements a model for the directories list.
+ */
 public class FileListModel extends AbstractListModel<String> {
+	
+	//----------------------------------------------
+	// Class constants
+	
 	/**
 	 * Eclipse generated for serialization
 	 */
 	private static final long serialVersionUID = 6560570616700652378L;
+	
+	//----------------------------------------------
+	// Class properties
+	
+	// Data for the model
 	private List<String> model = null;
 
+	//----------------------------------------------
+	// Class methods
+	
 	public FileListModel() {
 		this.model = new ArrayList<String>();
 	}
@@ -24,14 +39,22 @@ public class FileListModel extends AbstractListModel<String> {
 		return this.model.size();
 	}
 
+	/**
+	 * This method adds an entry to the model	
+	 * @param directory New directory to appear in the model
+	 */
 	public void addEntry(String directory) {
 		this.model.add(directory);
 		Collections.sort(this.model);
 		fireContentsChanged(this, 0, this.model.size());
 	}
 
-	public void removeEntry(int índice) {
-		this.model.remove(índice);
+	/**
+	 * This method removes an entry from the model
+	 * @param index Index of the entry (0 <= index < model length) 
+	 */
+	public void removeEntry(int index) {
+		this.model.remove(index);
 		fireContentsChanged(this, 0, this.model.size());
 	}
 }
