+++ ClassFinder/src/main/java/jars/search/Main.java
@@ -5,12 +5,15 @@ import javax.swing.JPanel;
 import javax.swing.UIManager;
 
 public class Main {
+	
 	public static void main(String[] args) {
+		
 		try {
 			UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
+		
 		JFrame principal = new JFrame("Class Finder App");
 
 		JPanel panel = new SearchPanel();
