+++ ClassFinder/src/main/java/jars/search/SearchListener.java
@@ -3,10 +3,20 @@ package jars.search;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+/**
+ * This class implements an action listener for the "Search" button.
+ */
 final class SearchListener implements ActionListener {
-
+	
+	//-------------------------------------------------------
+	// Class properties 
+	
+	// Reference to the search panel
 	private SearchPanel panel;
 
+	//-------------------------------------------------------
+	// Class properties 
+	
 	public SearchListener(SearchPanel searchPanel) {
 		panel = searchPanel;
 	}
